Add ProcessLauncherWidget to SessionSetupDialog
Add UI for upcoming feature to allow profiling from entry
point. This adds a widget to specify an executable to launch,
arguments and a working directory, as well as a checkbox to
control whether we should pause the launched process at entry
point.

https://screenshot.googleplex.com/95pLBxXXL7yYCAA.png

The widget is only visible when launching Orbit with
"--local" and "--devmode".