Implement MouseEvent in CaptureViewElements
In this PR we are implementing mouse events for CaptureViewElements.
This events will unify all mouse interactions (currently mixed by picking +
custom implementations).

For now we are using it for MouseMove events, MouseWheel and
LeftUp/RightUp. Picking is the harder case since Timers will still use the
PickingManager and will come in a future PR.

- We are fixing highlighting of the scrollbars (http://b/231533802).
- We are also adding a few missing conversions from ScreenToWorld in the Slider.

Bug: http://b/230441102.

Test: Load a capture, check sliders works as expected with MouseMoved
events.