Refactor symbol loading lists in OrbitApp (#3672)
* Refactor symbol file loading lists

Beforehand there was only symbol_files_currently_retrieved_ which
guarded the complete retrieval process (finding local symbols and if not
found downloading from remote). symbol_files_currently_retrieved_ also
stored the StopSource for downloads. This commit adds the list
symbol_files_currently_downloading_ which only contains the modules that
are actually downloaded (not local finding).
symbol_files_currently_retrieved_ still guards the whole retrieval
process but the StopSource saving is moved to the downloading list.

Bug: part of http://b/231132857

* Use symbols_currently_loading_ differently

Beforehand symbols_currently_loading_ was the list of modules for which
symbols are currently loaded into the module (extracted from file) and
basically guarded the function LoadSymbols(). Afterwards
symbols_currently_loading_ is the list of modules for which symbol
loading overall is in progress, guarding RetrieveModuleAndLoadSymbols().
LoadSymbols() is still guarded, because it is only called from
RetrieveModuleAndLoadSymbols().

This also removes the need for an additional check (removed code).
Removing this check is in line with the optimal solution discussed in
http://b/205002963

Bug: part of http://b/231132857