Remove CHECK that causes PE/COFF unwinding to fail in name resolution
We remove a CHECK that gets triggered during unwinding when resolving
function names with Object::GetFunctionName(). Other CHECks for similar
methods (GetSoname(), GetBuildID(), GetGlobalVariableOffset()) are left
in, as these methods are not used during unwinding as used by Orbit.

This is a workaround for now and not the ideal solution. However, we
don't consider it reasonable to implement GetFunctionName() right now
or make other changes due to the following reasons:
* Resolving names can be turned off in the unwinder, however we
want to wait until we have automatic symbol loading implemented on the
client. Hence we need GetFunctionName() for ELF modules to still be
called and the same interface is used for PE/COFF modules.
* A proper implementation of GetFunctionName() for PE/COFF would need
to parse PDB files (which typically would not even be available
remotely), which is either a lot of work or requires an additional
dependency.

That said, we could partially resolve names for Wine dlls (using
.debug_info). This is worth considering as a follow-up step.

Tested: Unit tests.
Bug: http://b/194768602