Display Symbol loading status in the module list (#3676)
* Add symbol loading methods to AppInterface

This adds IsModuleDownloading and GetSymbolLoadingStateForModule to
orbit_data_views::AppInterface. GetSymbolLoadingStateForModule is used
to display the state of symbol loading in the module list.
IsModuleDownloading will in an upcoming PR be used directly to determine
whether to show "Stop download..." in the context menu.

* Add ModuleDataView symbol loading status

This changes the ModuleDataView to display the symbol loading status for
each module in the column "Symbols" (renamed from "Loaded") and adds
tests. This is for now hidden behind the devmode flag

Bug: http://b/231132857