Add Stop Download to context menu
This adds the "Stop Downloading..." action to the context menu of the
ModuleListView. There are 4 different modes the context menu can be in.
1. Only current downloads are selected, this results in the
"Stop Download..." action being visible and the "Load Symbols" action
invisible. 2. No download is selected, but modules that can be loaded,
this results in only the "Load Symbols" action being visible. 3. No
download is selected and no loadable module, this results in the "Load
Symbols" action being shown, but disabled. 4. Downloads and loadable
modules are selected, this results in both actions being shown.

This is hidden behind the devmode flag (bool new_ui_ in ModulesDataView)

Test: Manual test with devmode on and off
Bug: http://b/231578940