Manual instrumentation on Windows
The core addition is "EnableInTraceeWindows.cpp" which injects
OrbitApi.dll in a target process if it was not already loaded. We then
use our "CreateRemoteThread" function to enable/disable the Api on
capture Start/Stop.

https://screenshot.googleplex.com/9z2kNJfKCepRxaW

Note that orbit.dll was renamed OrbitApi.dll to avoid naming conflict
with our .pdb files for Orbit.exe and orbit.dll.

Tests: Profiled OrbitTest with Api tracing enabled on Windows and Linux.