Do not discard kernel samples with frame pointers
We used to discard samples in kernel code (triggered by user code),
as this was also going to discard samples falling into the 'int3'
of a uprobe.

However, later we added detection of the uprobe module. And now
we are missing samples which changes the distribution of samples.

In this change, we re-enable samples on kernel code. In most cases
we will receive completely valid samples. For the uprobes case,
we are still able to mark them as unwinding error, which also
has the benefit of exposing them to the user.

Test: Carefully inspected samples on "uprobes_target"
Bug: http://b/231672823