Fix bottom-up E2E test (#3682)
To support frame pointer unwinding in the bottom up test, we
load the symbols of libc. This yields to the fact that "ioctl"
is actually called "__ioctl". The test failed here.

Also, the test was missing to load the symbols of libdrm.so.

Test: Run E2E test locally.
Bug: http://b/231672823

Note, this does not fix the flakyness by frame pointer sampling
discarding kernel samples. This gets fixed by #3681.