Fixes to horizontal zoom & tests (#3678)
Fixes the issue that zooming when already at maximum zoom level
started to move the time window.

To reproduce: Zoom in until at max zoom level. Trying to zoom further
will start moving the time window slightly to the left or right, depending
on the position of the mouse.

Also fixes some smaller inconsistencies in the tests.