Support anonymously mapped PE/COFF files in MapInfo::GetObject (#3590)
While unwinding, if a program counter falls inside an anonymous executable
mapping, we need to detect whether this mapping actually corresponds to the
.text section of a PE, so that the necessary unwind information can be
retrieved. This happens if the offset of the section in the file
(`PointerToRawData`) is not congruent to the offset when loaded into memory
(`VirtualAddress`) modulo the page size, and Wine had no option other than
creating an anonymous mapping and copying the content of the section into it.

The detection logic, implemented by
`MapInfo::GetFileMemoryFromAnonExecMapIfPeCoffTextSection`, is the following:
Consider the previous file mapping, and consider the "group" of mappings for
this file. If this is a PE, and there is no executable file mapping for this PE
(in this "group" of mappings), and the address range where the .text section of
this PE should be mapped (based on the address of the first mapping in the
"group" and on the .text section's `VirtualAddress` and `VirtualSize`) is
contained in the anonymous mapping's address range, then we can be fairly sure
that the anonymous mapping does indeed correspond to the .text of the PE.

This logic is conceptually the same as the one implemented in
`orbit_object_utils::ParseMaps` by https://github.com/google/orbit/pull/3568.

More information on the problem in the documents linked in the bug.

NOTE: This adds support for `PeCoff` to `MapInfo::GetObject`, whether from an
anonymous executable map or from a regular file mapping. So this effectively
enables mixed callstack unwinding.

Bug: http://b/226562022

Test:
- Unit tests.
- Tried as part of the mixed callstack unwinding prototype on unaligned
  `triangle.exe` and on `Virtus`: with this, unwinding also proceeds through
  frames of the unaligned binaries, leaving close to no unwind errors.