Change the target label to no longer be a corner widget (#3688)
This was causing issues with accessibility and E2E tests as the corner
widget did not show up in the accessibilit tree. With this, the E2E
tests can now verify the connection to the Stadia instance by checking
the target label.

This also fixes the weird "flickering" of the menu bar that occured
during update of the CPU %.

Bug: b/230690933
Tests: Manual testing, and verified the accessibility tree using
Accessibility Insights.