Fix crash on Add Iterator click when MS or MA selected (#3691)
When a user selected a number of dynamically instrumented
functions and a number of manually instrumented scopes,
the "Add Iterator" action is active. A click on was crashing
the client.

The fix consists in a check for whether the scope_id corresponds
to dynamic instrumentation. The same thing is now done for
"Add Frametrack" aciton. These checks are followed by
ORBIT_CHECK that ensures an InstrumentedFunction instance
is found.

This also adds `ScopeDataView` abstract class, which all the 
`DataView`s that handle scopes should inherit from. Frametrack
enable/disable functionality is moved to ScopeDataView from 
DataView. 

Tests: Manual, Unit
Bug: http://b/232058826