Remove flag files for e2e tests.
Previously an e2e test script was only run if a flag file of the same
name was present in OrbitDirectory\automation_tests\flags directory.
This mechanism was intended to be used to deactive a test with a
single, trivial change in the github repo.

We never used this so this is only clutter and an unnecessary
complication.

Bug: http://b/208997475