Fixup: Modify MouseEvent
This Fixup address some comments about public and private order of mouse
events and also change the way to address the special case of MouseMove.
Now, an element will react to MouseEvents either if the mouse is over
the element or is a MouseMove event to switch the state of
is_mouse_over.

In addition we split IsMouseOver() with ShouldReactToMouseOver() and we
fix a hidden issue in ContainsPoint (it shouldn't count the border
Pos()+Size()).

We also add unit tests for TimeGraph, Slider and CaptureViewElement
test classes.

TODO: I'm not 100% sure about having mouse position in MouseWheel
events.