Only send "AddressInfos" once
While in the ProducerEventProcessor, we already hash
the strings of AddressInfo events and only send them once,
we are still sending up to 256 (4x64) bytes PER frame, for
each callstack. Even if we've already sent the data
for a respective absolute address.

This simple change, only sends the address infos once per
absolute address.

On a 18sec capture of OrbitService, this already reduced
the capture size from 3MB to 1.5MB.

Test: Capture OrbitService and validate symbols + Unit tests
Bug: http://b/230315380