Make text eliding consistent
The text eliding in table views is different to what we use in the
top-down and bottom-up views.

Users have reported that they like right-elided text more over the
middle-elided one.

This change changes the table views to right elided text - as it is in
the tree views (bottom up and top down).

Bug: http://b/197498131