Make a temporal solution to fix orbit_tracks E2E test
This PR is making a hack to temporarily solve orbit_tracks E2E test. The
issue is that SelectTrack is producing events out of order:
(http://b/233028574).

So, the hack is simply not having a SelectTrack followed by a
MoveTrack because we have a weird state with 2 picked tracks where
moving tracks doesn't work as expected.

Bug: http://b/232887653

Test: E2E runs.