Flip time at the cursor when needed
When the mouse was close to the right border of the capture window, the last
digits of the "time at cursor" indicator were cut of.

In this PR we are flipping only the mouse_label when there is no space
at the right but instead there is space to completely draw the label at
the left. This last check is only made for really tiny screens (< 100/150
width), where there could be no space at either side. In that case,  we still
prefer to draw it at the right.

http://screenshot/3DKBDSgnGNKH3Dg

Bug: b/231158495.

Test: Load a capture. Zoom-in/out. Move mouse to the corners.