Also show module names in CallstackThreadBar's tooltips
The recent work on Silenus showed that the additional information is very
useful, in particular when the function is in a module that the user doesn't
know about, or when there is no symbol for the function and only the absolute
address is shown.

For the frames in the callstack, I chose "module | function". This is similar to
Visual Studio's "module!function", but I think it's a bit more readable.

Screenshots:
http://screen/rvJx5qbpa9nsBfb
http://screen/BcNjjrQEuEUgkPV

Bug: http://b/233200089

Test: Various manual tests.