Detect PE text section based on characteristics, not name (#3723)
This brings `unwindstack::PeCoffInterface` in line with
`orbit_object_utils::CreateCoffFile`.
Section characteristics are more reliable because they determine how the section
is loaded in memory, while names are an aid for tools and can be changed as a
way of obfuscating the binary. For example, we have observed that in some cases
the section called ".text" is not executable and doesn't actually contain code.

Additionally, while the case of multiple executable sections is more complex,
this method reports the first executable section, which seems to be the text
section we are looking for in all observed cases. After all, that section is at
least where the code starts.

On a complex Silenus game, this dramatically reduces the number of unwind
errors, because which anonymous map corresponds to the text section can now be
detected correctly.

Also address a couple of clang-tidy warnings along the way.

Test: Updated unit tests. Tried on aforementioned game.