Add capture loading for Mizar
With this commits we begin the work on Mizar (aka comparison tool).
This adds abstractions to own CaptureData as well as
load it from a capture file. For now we load the sampling data and
instrumentation of type D and MS.

This also:
1. Renames and moves the earlier static function
`LoadCaptureFromNewFormat`.
2. Introduces AbstractCaptureListener implementing some of the
`CaptureListener` virtual methods, which `OrbitApp` and `MizarData`
inherit from.
3. Adds a dummy main function for tests and experiments.

Tests: Manual
Bug: http://b/233869757