Cache current pid value in GetCurrentProcessId() (#3728)
GetCurrentProcessId() is called on every introspection scope.
Profiling OrbitService, we saw that 2.6% of samples fell in
GetCurrentProcessId() on a test version of OrbitService that
spawns a thread that generates dummy introspection events. This
change brings the 2.6% number down to 0.1%. We now cache the
return value of getpid() in a static variable instead of calling
the function all the time. This change does not significantly
improve the performance of the Api scopes since similar caching was
allready done at the caller level.