Set grpc server-side maximum message size
We were setting the maximum grpc message size on the client side
but not on the server side. This caused an error when receiving
messages of size greater than 4194304.

Tests: select thousands of function to instrument and start a capture.
Before the change, we would get a popup with the message size error.