Extend keys for navigation in CaptureWindow
As it was planned in go/stadia-orbit-navigation, we are extending keys for
navigation similarly we did with mouse wheel events.

In this PR we added arrow keys to perform a scrolling or a pan if no timers
has been selected (this could be done nicer if we do something similar as we
did for mouse events: http://b/234116147). We also included PageUp/PageDown
and Ctrl + '+' and '-' for VerticalZoom.

Tests were adapted for the new commands. And a bit of a renaming methods had
been added to VerticalScrolling in TrackContainer.
Get/SetVerticalScrollingOffset shouldn't be public, but this implies a
refactor in panning that I'm not planning to do soon.

Also the toggle help was outdated. I'm opened to suggestions, especially for
the new "Zoom to specific time range" (http://screenshot.com/7JYhSijrBbHfg3o).

Bug: http://b/168101815

Test: Load a capture, check all the keyboard and mouse wheel new commands.