Make PeCoff::GetRelPc independent from the text section's VirtualAddress
Before, `PeCoff::GetRelPc` was assuming that the start of the map coincide with
the start of the executable section. For some games on Silenus we have seen that
this doesn't always hold, as a single executable section can end up being mapped
into multiple executable maps. `PeCoff::GetRelPc` can be made more generic to
support this, by applying the same logic as `Elf::GetRelPc`, which makes use of
`MapInfo::object_offset`. Compared to `Elf`, for PEs the offset in the file
isn't equal to the offset in memory relative to the load bias, so we need to
convert from file offset to RVA. For this, we add
`PeCoffInterface::MapFromFileOffsetToRva`.

Bug: http://b/234826808

Test:
- Update unit tests.
- Tried on complex Silenus game: the unwind errors targeted by this fix are
  gone.
- Tried on `triangle.exe`: unwinding still works correctly. This case is
  important because the alignment of `triangle.exe`'s executable section makes
  it so that Wine has to map it into an anonymous map.