Add capture loading for Mizar
With this commits we begin the work on Mizar (aka comparison tool).
This adds abstractions to own CaptureData as well as
load it from a capture file. For now we load the sampling data and
instrumentation of type D and MS.

This also adds a dummy main function for tests and experiments.

Tests: Manual, Unit
Bug: http://b/233869757