Relax check for timers in manual instrumentation test.
Some of the threads we filter for do not contain timers since they are
created by the lib Orbit injects into the process.
Strictly speaking we could test for exactly three of the threads
containing timers. Since the code is already there and we loose nothing
of value we just check for at least one of the track containing timers.
Details are in the bug.

Test: Local run.
Bug: b/229709116