Save which symbol loading resulted in an error
This adds a set to OrbitApp of all the modules where symbol loading
resulted in an error and uses that information in
`OrbitApp::GetSymbolLoadingStateForModule`.

Test: Start with --devmode, attempt to load symbols for a module where
symbols are missing. After dismissing the error, the module list
displays "error" (with auto symbol loading, there will not be an error
pop up anymore, but only the "error" in the list).
Bug: http://b/234581674