Add symbol loading to `MizarData`
As Mizar needs to work with Windows captures void
of `LinuxAddressInfo`s, it is necessary to add symbol
loading from symbol files.

This adds automatic loading of symbols from the folders
Orbit App is configured to search for them and from
the caches of Orbit app.

We will have to work with binaries with no build ids.
Therefore, Mizar resorts to file size check instead.
The corresponding functionality is added to `Symbolhelper`.

Tests: Manual, Unit
Bug: http://b/233856442