Scale consistently in TimeGraph
In this PR we are scaling consistently across all elements in TimeGraph, so:

- We are only having one font size (before was impossible because text in
the Timeline was overlapping).
- We are scaling symbol sizes in Buttons accordingly.
- Move a few hardcoded values from Slider to Layout, so they can scale.
- An additional ceiling needed for the horizontal scrollbar to avoid
  having this situation because OpenGl: http://screenshot/3mVGcjPYHgfYuKF

And in addition:
- Rename SliderWidth x SliderWide.
- Refactor a bit Slider class.

Some screenshots:
TimeGraph after zooming-in: http://screenshot/AvYgew2RL5ToNaU
Min size of Scrollbar after zooming-in: http://screenshot/kPY4FfmPBnpeppL

Bugs:
http://b/234358833 -> Scale Timeline - Button - Scrollbar
http://b/200905950 -> Text of iterators doesn't get zoomed.
http://b/230442062 -> Refactor Sliders (although I won't close it since
there is still a bit of cleaning work we can do there).

TODO:
http://b/177200179 -> Scale Qt part (not a priority for now).