Also use COFF symbol table (on top of DWARF) to load symbols from a PE (#3755)
Originally we were loading symbols from the COFF symbol table.
However, this doesn't contain function sizes, so
https://github.com/google/orbit/pull/2489 switched to using DWARF debug info.
But the DWARF debug info misses some functions compared to the COFF symbol
table.

We bring back the use of the COFF symbol table and integrate its information
with the DWARF debug info. Functions that are in the COFF symbol table but not
in the DWARF debug info are added to the symbol list, and their size is computed
as the distance from the next function. This can overestimate the size but is
better than not showing the function at all.

Note that this is mostly for Silenus, as MinGW adds the COFF symbol table and
the DWARF debug info to the DLLs. Other cases normally use PDB files.

Bug: http://b/234585733

Test:
- Update unit test.
- The missing symbols for Silenus's `ntdll.dll` now show up. For example
  `NtQueryPerformanceCounter` (which triggered this investigation), which also
  has its size set correctly and can now be shown in Orbit's disassembly.