Represent callstacks as sequences of sampled function ids
For now the callstack samples are represented as sequences of
addresses. And these addresses vary run-to-run. Mizar needs a
general way to treat these based on sampled function ids.

This renames MizarDataWithSampledFunctionIds to
MizarPairedData.

The method MizarPairedData::ForEachCallstackEvent is added.
One of its arguemnts is a callable `Action` that takes a
callstack that is represented as a sequence of
sampled function ids.

Also, this makes the methods CallstackData::ForEach*
take a templated Action, not std::function.

Bug: http://b/235197552
Tets: Unit