Fix crash in user space instrumentation.
If a function constains a jump into its first five bytes we must not instrument
it: During instrumentation we would overwrite these five bytes and therefore
any jump into this address range would just execute garbage.

Previously this issue was competely ignored. This PR provides a
heuristic catching the cases I found "in the wild". Some additional details are
documented in the bug.

Test: Runs with example games, unit tests in the PR.
Bug: http://b/234825095