Introspection thread states (#3747)
Send OrbitService's thread states when introspection is enabled so that
we have more context on what our own threads are doing when capturing.
SwitchesStatesNamesVisitor now filters multiple process ids.

Tests: Launch a capture with introspection enabled and see that we now have
thread states. Note that OrbitService threads that don't have instrumented scopes
are displayed as if they were part of the target process (b/234029571), this should
be fixed at some point, but is not trivial. This is not critical as the feature is internal.