Use CanceledOr in symbol loading (#3774)
With this commit the new type CanceledOr<T> is used in symbol loading.
This starts with ServiceDeployManager where CopyFileToLocal returns
`Future<ErrorMessageOr<CanceledOr<void>>>` and is passed down until
OrbitApp. In 2 cases (Preset loading and debug info loading, which is
used for disassembly & source code view), a cancellation is turned back
into an error. For regular symbol loading this change means that a
cancellation does not result in an ErrorMessage, which means a canceled
symbol loading result is not saved in
`modules_with_symbol_loading_error_`, which means its not displayed as
error in the module list.

Test: Manual. Start Orbit with --devmode, click "Load Symbols" on a
module, click "Stop Download" on that module. No error should appear.
(Beforehand an error with detailed message: "User canceled download"
was shown)
Bug: part of http://b/231579238