Consider non-"8-byte register pops" in PeCoffEpilog (#3771)
Before, we were assuming that the operand of a `pop` would always be a register,
but that's not the case. See for example https://www.felixcloutier.com/x86/pop.
That could fail a `CHECK`.

Also, https://docs.microsoft.com/en-us/cpp/build/prolog-and-epilog#epilog-code
prohibits `pop`s to registers that are not 8-byte in an epilog.

We split `HandlePopInstruction` into `ValidatePopInstruction` and
`HandleEightByteRegisterPopInstruction`. `ValidatePopInstruction` verifies that
we actually have an 8-byte register pop.

Bug: http://b/235458021

Test: Add unit tests. `detects_non_epilog_pop_to_memory` crashes without the
fix.