Use stop dialog (#3783)
* Add download_disabled_modules_ list

This adds download_disabled_modules_ to OrbitApp and uses it. The list
contains the modules for which the user checked "Never download symbols
for this module" before. Making this list persistent is done via
QSettingsStorageManager. A entry is removed from the list when the user
manually clicks "Load Symbols" on this module (LoadSymbolsManually).

Bug http://b/231579238

* Use StopSymbolDownloadDialog

This commit adds the usage of StopSymbolDownloadDialog to Orbit via the
method MainWindowInterface::DisplayStopDownloadDialog. When a user
clicks "Stop Download" in the module list context menu, the dialog is
shown.

Test: Manual, start with devmode, click Stop Download
Bug: http://b/231579238