Detect multiple anonymous executable PE maps in MapInfo::GetObject
https://github.com/google/orbit/pull/3590 added the detection of anonymous
executable maps that belong to PEs with a file alignment incompatible with
`mmap`: this is done in
`MapInfo::GetFileMemoryFromAnonExecMapIfPeCoffTextSection`. The detection could
be quite strict by assuming that a PE only has one executable section. We
observed that this doesn't always hold, so we need to relax some conditions. Now
we simply accept any anonymous executable map that lies inside the address range
of size `PeCoff::GetSizeOfImage()` that starts at the first address at which the
PE is mapped. This logic is similar to what is done in
`orbit_object_utils::ParseMaps` after https://github.com/google/orbit/pull/3786.

The start of these maps doesn't necessarily correspond to an offset in the file:
not only because an anonymous map doesn't have an offset by definition, but
especially because the start of the anonymous map could correspond to no byte in
the file.
But `MapInfo::object_offset()` is needed to compute the relative PC from a PC.
To support this case, we add `MapInfo::object_rva()` and make `PeCoff::GetRelPc`
choose what to use. In all other cases (file mappings, ELF files),
`MapInfo::object_rva()` is zero and `MapInfo::object_offset()` applies as
before.

Bug: http://b/235480245

Test:
- Adjusted `MapInfoGetObjectTest.cpp`.
- Tried on `triangle.exe`: already worked fine before, still works.
- Tried on complex Silenus game that highlighted the problem. Together with
  https://github.com/google/orbit/pull/3791, unwinding errors are basically gone
  (except for the ones related to `__wine_syscall_dispatcher`, as usual).