Reset PeCoffEpilog::last_error_ in DetectAndHandleEpilog (#3765)
`PeCoffEpilog::GetLastError` is used to tell whether
`PeCoffEpilog::DetectAndHandleEpilog` returning false means that we are not in
the epilog or that an error occurred. This means that
`PeCoffEpilog::last_error_` needs to be reset for every invocation of
`PeCoffEpilog::DetectAndHandleEpilog`. This didn't happen before, always keeping
`PeCoffEpilog` in an error state.

Test: Add unit test: failed before, passes now.