PeCoffEpilog now supports multiple executable sections
With PEs with multiple executable sections, when a program counter was inside an
executable section other than the first, the computation of the offset in the
file was producing the wrong value. This was making us to read bytes from the
wrong location, hence causing an error in epilog detection because
disassemblying would also most likely fail.

We now support multiple executable sections by following the same pattern as in
other parts of PE unwinding with the `MapFromRvaToFileOffset` function.

Bug: http://b/236091987

Test:
- Add and adjust unit tests.
- Tried on complex Silenus games that highlighted the problem. Together with
  another upcoming fix that handles multiple executable sections mapped
  anonymously, this drastically reduces unwind errors (leaving basically only
  `__wine_syscall_dispatcher`).