Expose flags when allocating memory in tracee.
When calling mmap in the tracee we previously always allocated the
memory with flags MAP_PRIVATE | MAP_ANONYMOUS.
Now we'll also need to allocate some memory for a stack of a new thread
(context: see bug blow). The flags required here for that are:
MAP_ANONYMOUS | MAP_PRIVATE | MAP_STACK | MAP_GROWSDOWN.
We'll keep the MAP_PRIVATE | MAP_ANONYMOUS as a default but expose the
flags to facilitate that use case.

Bug: b/236121842