Fix error handling when detaching from a process.
The target process might spawn additional threads while we are attached
to it. When we detach from the process detaching from these threads
fails (since we never attached to that thread). We can simply ignore
these errors.

We will need this when we create an additional threads inside the
process. Context is here: b/236121842