`ModuleUpdateEvent`s consider multiple executable sections/maps
`UprobesUnwindingVisitor` uses `PERF_RECORD_MMAP` events to keep
`unwindstack::Maps` up to date for unwinding, but also to send
`ModuleUpdateEvent`s to the client. Before, it was assuming a one-to-one
relationship between executable maps and modules. In particular for games
running under Wine, we have seen that this is sometimes not the case, both for
ELF files (even with a single executable loadable segment) and for PEs (and PEs
can also simply have multiple executable sections).

Now, we try to send `ModuleUpdateEvent`s with an address range that covers all
the executable mappings for the same file, similarly to
https://github.com/google/orbit/pull/3786. In order to do this, we have to also
look at the existing maps around the new one.

And as usual, for PEs we also have to consider executable sections mapped into
anonymous executable maps, for which we take advantage of the existing logic in
`unwindstack::MapInfo::GetFileMemoryFromAnonExecMapIfPeCoffTextSection`,
as was already the case after https://github.com/google/orbit/pull/3667.

Bug: http://b/235481314
Bug: http://b/235480245

Test:
- Extend `UprobesUnwindingVisitorMmapTest`.
- Tried by capturing unaligned `triangle.exe` from the beginning by setting
  `LD_PRELOAD=libdebugger_preload.so`. Verified that the callstacks make sense.