ParseMaps handles multiple executable sections and/or maps (#3786)
We are solving the issue with object files that are mapped into memory into
multiple mappings. This can happen if the single executable section is split
across multiple mappings, or if there is more than one executable section (the
latter should only be the case for PEs).

Before, each executable mapping was seen by Orbit as a separate module. In these
cases, the correspondence between symbols and addresses was computed
incorrectly, because only one of these "modules" had the correct address in
memory at which the first executable section of the object file was mapped.

Now, when scanning `/proc/[pid]/maps` in `orbit_object_utils::ParseMaps`, we
only create a single module for a sequence of mappings that correspond to the
same file. In particular, we consider an address range that covers all the
executable mappings for that file. This could also cover non-executable
sections, but it doesn't matter. The only assumption is that the beginning of
this address range still corresponds to the "executable segment offset", i.e.,
the "executable segment offset" corresponds to the first executable section.

We need to relax the conditions applied when detecting anonymous executable maps
that corresponds to PEs with a file alignment incompatible with `mmap`. Now we
simply accept any anonymous executable map that lies inside the address range of
size `ObjectFile::GetImageSize()` that starts at the first address at which the
PE is mapped.

Bug: http://b/235481314
Bug: http://b/235480245

Test:
- Adjusted and expanded `LinuxMapTest.cpp`.
- Tried on `triangle.exe`: already worked fine before, still works.
- Tried on complex Silenus games: now each module only shows up once and symbols
  are correctly assigned to addresses.