Ignore instrumentation events from threads spawned by injected lib.
Directly after the injection of the library we create a new thread using the
clone syscall and execute the initialization code in this thread.

We resume the execution of the game and allow the initlization to happen. After
that we detect the newly spawned threads, attach again to the target and set
the tids in the lib so that we can ignore these threads in EntryPayload.

The bug contains some additional context.

Bug: b/236121842
Test: Covered by existing unit tests. Also manual testing with several
targets.