Receive "functions to stop at" in LinuxTracing
Adds the functions to stop unwinding at to the capture options proto,
reads them and passes them to the unwinder and
UprobesUnwindingVisitor.

Test: E2E Integration and capture with __wine_syscall_dispatcher
Bug: http://b/236121587

Receive "functions to stop at" in LinuxTracing

Adds the functions to stop unwinding at to the capture options proto,
reads them and passes them to the unwinder and
UprobesUnwindingVisitor.

Test: E2E Integration and capture with __wine_syscall_dispatcher
Bug: http://b/236121587