Do not apply majority filtering to functions to stop at
We may receive callstacks that are marked as "complete", but do not
end in the expected function (start or clone). Therefore, we
implemented a mechanism on the client to filter those callstacks
based on the majority.

Now that we have functions were we request the unwinder to stop at
(__wine_syscall_dispatcher), callstacks ending in those functions to
stop at might either get filtered out, or make the majority (and thus
clone would get filtered out).

With this change, we are excluding the functions to stop unwinding
at, from the calculation.

Bug: http://b/236121587
Test: Unit Test