Improve StartNewCapture.
Previously the timestamp was obtained from within the target process.
This is not needed (can be done in OrbitService) and potentially
dangerous since it is done in a halted thread and it is not entirely
clear if the call is reentrant.

Since StartNewCapture is not spawning any new threads anymore we can
remove the call to AttachAndStopNewThreadsOfProcess.

Bug: b/232684803
Test: Unit tests plus manual runs.