Don't expect symbol loading errors in E2E tests
In this PR we are modifying E2E tests to not expect error boxes when
loading symbols. This is a change that will appear in the next release
as part of the automatic symbol loading feature.

Instead we will check a weaker condition: "At least one module has been
loaded".

Bug: http://b/236928772

Test: Symbol cache E2E is passing locally (in main).