Add ProcessData::FindModulesByFilename
We will need to know the absolute addresses of certain functions
we want to stop unwinding at (in particular __wine_syscall_dispatcher).

As `FunctionInfo::GetAbsoluteAddress` is deprecated since it does not
cover loading the exact same module twice, we need a way to retrive
the module start addresses from the process.

This change adds a `FindModulesByFileName` function to the process.

Test: Unit test
Bug: http://b/236121587