Stop unwinding at __wine_syscall_dispatcher
This sends `__wine_syscall_dispatcher` (if loaded by the process) to
the service as a function to stop unwinding at and therefore omits
unwinding errors related to this function.

As the detection of `__wine_syscall_dispatcher` relies on the symbols
for `ntdll.so` being loaded, we prioritize loading of this module.
This also requires auto-symbol loading to be affective.

Follow-up:
Implement a solution that actually allows unwinding through the
dispatcher into the user-space callstack.

Test: Inspect callstacks when profiling on Silenus
Bug: http://b/236121587