Adapt loading symbols in E2E tests to just wait for them (#3824)
* Adapt loading symbols in E2E tests to just wait for them

In this PR we are adapting E2E tests to instead of loading symbols
manually, just wait until all the symbols are loaded automatically.

We still will need LoadSymbol method for orbit_symbol_locations E2E
tests, since they won't be able to be loaded automatically.

After thinking for a while I decided to leave 2 different methods:

- WaitForLoadingSymbolsAndVerifyCache() that it's doing exactly the same
  than the previously LoadAllSymbolsAndVerifyCache() but without the
  manual loading action (used for cache and measuring the time).
- WaitForLoadingSymbolsAndCheckModule() that it's doing the proper
  replacement from LoadSymbols and it's waiting and after checking that
  the module was loaded.

I checked a few E2E test locally and found an issue with the
requirements now that we load all the symbols because there was more
than 1 function that matched with VkQueuePresent, so I relaxed the
requirements for VerifySymbolLoaded to check "at least" one row.

This PR is a bit of draft. I would like to have opinions but I need to
- Wait for merging https://github.com/google/orbit/pull/3814 and
  https://github.com/google/orbit/pull/3823.
- Rebase to that PRs.
- Create a signed build and run E2E tests to check that everything is
  working.
- Recalculate.

I expect to have an issue in symbol_locations (I didn't test that one
locally yet). Some test could also fail because loading symbols after
a capture. In that case I will add a wait before each Capture taken.

TODO: Align with Anton if we need to additionaly test post-symbol
loading and how.

Bug: http://b/236703806
Test:
- symbol_cache E2E pass locally.
- add_iterator E2E pass locally.
- tracks E2E fails because devmode has more tracks. It should pass with
the changing of flags.
- capture_repeatedly E2E fails because of http://b/236976315.
- I didn't run symbol_locations. I will try to run it locally tomorrow.

* Fixup: Comment + module

* Fixup: WaitForLoadingSymbols is symbol_locations.py