Throttle calls of UpdateAfterSymbolLoading (#3825)
This avoids that multiple subsequent calls to `UpdateAfterSymbolLoading`
block the UI and do the very same post processing over and over again.

The updates will be limited to at most once per second. If multiple
updates get requested during the same interval the will be combined and
only a single update will be executed after the second has passed.