Modify the preset tests to wait for symbol loading (#3841)
Add WaitForLoadingSymbolsAndVerifyCache to the preset tests.

This could use some slight cleanup in the future: We're now using
`WaitForLoadingSymbolsAndVerifyCache` everywhere, but this test is
doing more than it needs to do. We should split the "wait for symbol
loading" functionality so it can be used separately from verifying the
cache status, but I'd do this in a follow-up.