Check that Timeline exist in E2E tests
In this PR we are adding a check for existance for the Timeline in the
E2E tests. We are also checking and assuming that this timeline doesn't
have any children.

We are also modifying a bit how we get the TrackContainer so it doesn't
depend on the order of the children in TimeGraph.

Test: Run locally instrument_function and add_iterator E2E tests.
Bug: http://b/237096560