Revert E2E to check cache after waiting for symbols (#3849)
In https://github.com/google/orbit/pull/3824 we make most E2E tests wait
for symbols to be loaded instead of loading them. I decided to also
check that the modules are in the cache as well as part of waiting and
checking for symbols to be loaded. However this is wrong for
orbit_symbol_locations and orbit_source_code_view. It's not expected to
have that modules in the cache since symbols are loaded from a local
folder.

So in this PR we move the method of waiting for symbols to be loaded as
a free function and just run that method. We only will check that the
modules are in the cache for the orbit_symbols_cache E2E test.

Test: Run locally symbol_cache, symbol_location, source_code and
instrument_function E2E test.

Bug: http://b/237270712 and http://b/237270158.