Add a sleep before checking a function is hooked
We think that load_preset_2 end is flaky because it takes a bit of time
to update the function list because there are too many after automatic
symbol loading.

So, in this PR we are adding a 100ms delay before checking for the
hooked function. We are also improving the log message in the case that
the function is not found.

Bug: http://b/237513889.
Test: Run load_preset_2 manually.